requirements.txt
groq==0.4.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.0